Move Forty endpoints to api v1 namespace
