+++ app/controllers/api/v1/pto_reports_controller.rb
@@ -1,0 +1,7 @@
+module Api
+  module V1
+    class PtoReportsController < Api::V1Controller
+      expose(:pto_reports) { PtoReport.all }
+    end
+  end
+end
