+++ app/controllers/api/v1_controller.rb
@@ -1,18 +1,3 @@
 module Api
-  class V1Controller < ApplicationController
-    protect_from_forgery with: :null_session
-
-    before_action :validate_client_token
-
-    private
-
-    def client_token_valid?
-      client_token = request.headers['X-CLIENT-TOKEN']
-      Rails.application.secrets.client_token == client_token
-    end
-
-    def validate_client_token
-      head :not_found unless client_token_valid?
-    end
-  end
+  class V1Controller < ApiController; end
 end
