--- app/controllers/pto_reports_controller.rb
@@ -1,3 +1,0 @@
-class PtoReportsController < ApiController
-  expose(:pto_reports) { PtoReport.all }
-end
