+++ app/views/api/v1/work_weeks/show.json.jbuilder
